El Analisis Lexico no tiene problemas con los espacios
