+++ Bibliotecas/Pila.h
@@ -1,0 +1,64 @@
+TSimbolos *PPila, *QPila, *AuxPila, *NuevoPila;
+
+void push(const char* a)
+{
+	NuevoPila = (TSimbolos*)(malloc(sizeof(TSimbolos)));
+	NuevoPila->liga = NULL;
+	strcpy(NuevoPila->derivacion, a);
+	if(PPila == NULL)
+	{
+		PPila = NuevoPila;
+		QPila = PPila;
+	}
+	else
+	{
+		QPila->liga = NuevoPila;
+		QPila = QPila->liga;	
+	}
+}
+
+void pop()
+{
+	AuxPila = PPila;
+	if(AuxPila != NULL)
+	{
+		if( PPila->liga != NULL )
+		{	
+			while(AuxPila->liga->liga != NULL)
+				AuxPila = AuxPila->liga;
+			QPila = AuxPila;
+			AuxPila = QPila->liga;
+			QPila->liga = NULL;
+			free(AuxPila);
+		}
+		else
+		{
+			free(AuxPila);
+			PPila = NULL;
+			QPila = NULL;
+		}	
+	}
+}
+
+void top()
+{
+	printf("%s\n", QPila->derivacion);
+}
+
+void ImprimeDerivacion()
+{
+	printf("=> ");
+	AuxPila = PPila;
+	while(AuxPila != NULL)
+	{
+		printf("%s", AuxPila->derivacion);
+		AuxPila = AuxPila->liga;
+	}
+	printf("\n");
+}
+
+void VaciaPila()
+{
+	while( PPila != NULL )
+		pop();
+}
