+++ archivofuente.txt
@@ -1,7 +1,8 @@
-int a, b, c;
+int a, b;
+a = 1;
+b = 5;
 
-while( a == 5 )
+while( a <= b )
 {
-        y = 8;
-        a = 4;
+    a = a + 1;
 }
